Connector: don't connect during construction.
The master_proxy is considered to be in 'connected' state if its
connector object is in 'connected' state. If this connector object
(which is created by the master_proxy constructors) starts connected,
then the master_proxy::connect() function assumes that it is not
necessary to send an OpenPDU to the master. That means that no session
will be established.

We circumvent this problem by letting the connector start in
'disconnected' state. Connection is established by the
master_proxy::connect() function.