Handle callbacks on destroyed timeout_timer's.
The time_timer class invokes async_wait() on its
boost::asio::deadline_timer object. If the timeout_timer object is
destroyed (and consequently, its deadline_timer) before the timer's
callback was invoked, the callback is invoked by asio nevertheless. This
led to a call of a member function of a destroyed object and resulted in
crashes.

Now, the callback is a static member function and can be called anytime.
Further, a static class member is introduced to track all existent
timeout_timer objects, so that the callback can identify when it is
invoked for a destroyed object.