Break up the concepts of 'value' and 'variable'.
This is a big design change. The 'variable' types were used to implement
two distinct concepts:
- Represent a value, e.g. an Integer value
- Represent a variable (an object performing actions on Set and Get
  requests).
The variable had an internal value. Simply returning this value is not
enough for a Get request. Conceptionally, a variable has no internal
state, but determines the current value on each Get request. Similarly,
on a Set request, the new value must effect something, storing it within
the variable object isn't enough. Thus, the internal value could be seen
as a cache.

This commit breaks the concepts an introduce two new classes (or rather
class hierarchies): 'value' and 'variable'.

The types Integer, Octet_String, Counter32 and so on are not longer
derived from 'variable', but from the new class 'value' (i.e. 'variable'
was basically renamed to 'value'). This class represents a value and is
not used to implement SNMP variables.

In addition, a new class 'variable' was introduced to represent the
variables. This class is derived by the newly-introduced
primitive_variable class, which is new the base class for
user-implemented SNMP variables.

The design (and code) still needs cleanup, and the documentation is now
heavily out-of-sync. This will be fixed in the up-coming commits.