ResponsePDU: Change handling of error_t values.
The ResponsePDU::set_error() function doesn't longer check its
parameters, and the ResponsePDU parse constructor now checks for all
values (to detect a malformed PDU). Note that the constructor
did accidentally only accept a subset of error_t values before.

This also fixes an error which appeared when performing SET to a
read-only primitive_variable: no response was sent in this case, because
ResponsePDU didn't accept the noAccess error (which means "read-only").