Merge *Value into *Variable.
After rethinking the design, I think that distinguishing between values
and variables makes no sense. E.g. IntegerValue is a class representing
an SNMP Integer "value", while IntegerVariable represents an SNMP Integer
variable. In a sense, IntegerValue is the type of the variable, while
the class IntegerVariable is - well, the type of the variable.

I decided to merge IntegerValue into IntegerVariable. I removed
the AbstractValue and Value<> classes. The IntegerVariable class now
inherits from AbstractVariable and is not abstract, i.e. it can be used
as SNMP variable without deriving a new class from it (for simple
cases).

The same is of course true for the other *Value / *Variable classes.